refactor: refactor pp results Signed-off-by: siarhei <Siarhei.Matsiushonak_ext@50hertz.com>
