+++ packages/contingency_analysis_pkg/src/toop_engine_contingency_analysis/pandapower/__init__.py
@@ -19,11 +19,11 @@ from toop_engine_contingency_analysis.pandapower.pandapower_helpers import (
     extract_contingencies_with_unique_pandapower_id,
     extract_monitored_elements_with_cgmes_id,
     extract_monitored_elements_with_unique_pandapower_id,
-    get_branch_results,
+    get_branch_results_polars,
     get_convergence_df,
     get_failed_va_diff_results,
     get_full_nminus1_definition_pandapower,
-    get_node_result_df,
+    get_node_results_polars,
     get_regulating_element_results,
     get_va_diff_results,
     translate_contingencies,
@@ -40,11 +40,11 @@ __all__ = [
     "extract_contingencies_with_unique_pandapower_id",
     "extract_monitored_elements_with_cgmes_id",
     "extract_monitored_elements_with_unique_pandapower_id",
-    "get_branch_results",
+    "get_branch_results_polars",
     "get_convergence_df",
     "get_failed_va_diff_results",
     "get_full_nminus1_definition_pandapower",
-    "get_node_result_df",
+    "get_node_results_polars",
     "get_regulating_element_results",
     "get_va_diff_results",
     "run_contingency_analysis_pandapower",
