+++ packages/contingency_analysis_pkg/src/toop_engine_contingency_analysis/pandapower/pandapower_helpers/__init__.py
@@ -16,10 +16,10 @@ from .extractors import (
 from .result_builders import get_convergence_df, get_full_nminus1_definition_pandapower
 from .results import (
     create_closed_bb_switches_graph,
-    get_branch_results,
+    get_branch_results_polars,
     get_failed_switch_results,
     get_failed_va_diff_results,
-    get_node_result_df,
+    get_node_results_polars,
     get_regulating_element_results,
     get_switch_mapped_elements,
     get_switch_results,
